// Capstone Project/src/app/pages/admin/admin-bank-account-management-page/screen/types.ts
export enum adminPageBankAccountEnum {
  GET_DATA_BANK_ACCOUNT_REQUEST = 'APP/ADMIN_PAGE/GET_DATA_BANK_ACCOUNT_REQUEST',
  GET_DATA_BANK_ACCOUNT_SUCCESS = 'APP/ADMIN_PAGE/GET_DATA_BANK_ACCOUNT_SUCCESS',
  GET_DATA_BANK_ACCOUNT_FAILURE = 'APP/ADMIN_PAGE/GET_DATA_BANK_ACCOUNT_FAILURE',

  UPDATE_DATA_BANK_ACCOUNT_REQUEST = 'APP/ADMIN_PAGE/UPDATE_DATA_BANK_ACCOUNT_REQUEST',
  UPDATE_DATA_BANK_ACCOUNT_SUCCESS = 'APP/ADMIN_PAGE/UPDATE_DATA_BANK_ACCOUNT_SUCCESS',
  UPDATE_DATA_BANK_ACCOUNT_FAILURE = 'APP/ADMIN_PAGE/UPDATE_DATA_BANK_ACCOUNT_FAILURE',

  CREATE_DATA_BANK_ACCOUNT_REQUEST = 'APP/ADMIN_PAGE/CREATE_DATA_BANK_ACCOUNT_REQUEST',
  CREATE_DATA_BANK_ACCOUNT_SUCCESS = 'APP/ADMIN_PAGE/CREATE_DATA_BANK_ACCOUNT_SUCCESS',
  CREATE_DATA_BANK_ACCOUNT_FAILURE = 'APP/ADMIN_PAGE/CREATE_DATA_BANK_ACCOUNT_FAILURE',

  DELETE_DATA_BANK_ACCOUNT_REQUEST = 'APP/ADMIN_PAGE/DELETE_DATA_BANK_ACCOUNT_REQUEST',
  DELETE_DATA_BANK_ACCOUNT_SUCCESS = 'APP/ADMIN_PAGE/DELETE_DATA_BANK_ACCOUNT_SUCCESS',
  DELETE_DATA_BANK_ACCOUNT_FAILURE = 'APP/ADMIN_PAGE/DELETE_DATA_BANK_ACCOUNT_FAILURE',

  CLEAR_DATA_BANK_ACCOUNT = 'APP/ADMIN_PAGE/CLEAR_DATA_BANK_ACCOUNT',
}